renamed to RoomEntrance, makes more sense
