+++ Assets/Prefabs/SceneManaging/RoomEntrance.prefab.meta
@@ -1,0 +1,7 @@
+fileFormatVersion: 2
+guid: 6b0d7f959f3251b419b65532668d6b47
+PrefabImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
