+++ Assets/Prefabs/SceneManaging/RoomEntranceSpawner.prefab.meta
@@ -1,0 +1,7 @@
+fileFormatVersion: 2
+guid: f007594aa708ace4a99ecc17408974bd
+PrefabImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
