+++ Assets/Scenes/Examples/SceneManaging/Inside.unity
@@ -542,6 +542,10 @@ PrefabInstance:
       propertyPath: correspondingExitName
       value: InsideEntrance
       objectReference: {fileID: 0}
+    - target: {fileID: 5143719055848018831, guid: 6b0d7f959f3251b419b65532668d6b47, type: 3}
+      propertyPath: correspondingEntranceName
+      value: InsideEntrance
+      objectReference: {fileID: 0}
     - target: {fileID: 7981659705172534632, guid: 6b0d7f959f3251b419b65532668d6b47, type: 3}
       propertyPath: m_RootOrder
       value: 1
