+++ Assets/Scenes/Examples/SceneManaging/Outside2.unity
@@ -179,6 +179,10 @@ PrefabInstance:
       propertyPath: correspondingExitName
       value: Outside2Entrance
       objectReference: {fileID: 0}
+    - target: {fileID: 5143719055848018831, guid: 6b0d7f959f3251b419b65532668d6b47, type: 3}
+      propertyPath: correspondingEntranceName
+      value: Ouside2Entrance
+      objectReference: {fileID: 0}
     - target: {fileID: 7981659705172534632, guid: 6b0d7f959f3251b419b65532668d6b47, type: 3}
       propertyPath: m_RootOrder
       value: 2
