+++ Assets/Scripts/SceneManaging/RoomEntranceController.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: e8ee2e25c43226d49b4e3c3eb849f238
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
