+++ Assets/Scripts/SceneManaging/RoomEntranceSpawner.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: d48764a3aec1f774b974cb7dc2749f21
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
