+++ Assets/Unity.VisualScripting.Generated.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: e5ea1a3376692d846b109841c1871a1e
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
