+++ Assets/Unity.VisualScripting.Generated/VisualScripting.Flow.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: ad22c623587c3a841aa108fc419a0cd0
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
