+++ Assets/Unity.VisualScripting.Generated/VisualScripting.Flow/UnitOptions.db.meta
@@ -1,0 +1,7 @@
+fileFormatVersion: 2
+guid: 86051824f80df2c4987aa48246c9d0cb
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
